docs/rag-tool/requirements.txt
fastapi
pandas
chromadb
python-dotenv
pyarrow
fastparquet